change the file structure
