+++ Solution/0001_TwoSum.cpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "Solution.h"
+#include "../Solution.h"
 
 using namespace std;
